Fixing the overflow in update interval counters

+++ Firmware/LilyGo/settings.h
@@ -26,13 +26,13 @@ typedef struct lilygoSettings {
   String weatherURL;
   String weatherUnit;
   String weatherLang;
-  int8_t weatherUpdateInterval;
+  int weatherUpdateInterval;
   // NTP Settings
   String ntpServer;
   int dstOffset;
   // location settings
   String geoipURL;
-  int8_t locationUpdateInterval;
+  int locationUpdateInterval;
   // WiFi settings
   String wifiApSSID;
   String wifiSSID;
@@ -43,7 +43,7 @@ typedef struct lilygoSettings {
   // General settings
   uint32_t loraMagic;
   bool displayBGrndBlack;
-  int16_t stepGoal;
+  int stepGoal;
   // Google settings
   String googleApiKey;
   String googleGeoLocationURL;
